requirements.txt
streamlit
torch
transformers
google-generativeai
scikit-learn
pandas
numpy